repo: rename /resources => /endpoints
